requirements.txt
# requirements.txt ACTUALIZADO
kivy==2.3.0
kivymd==1.2.0
requests==2.31.0
firebase-admin==6.2.0
python-dateutil==2.8.2
unidecode==1.3.8
certifi==2024.8.30
chardet==5.2.0
urllib3==2.2.2
python-dotenv==1.0.0  